refactor: move notion-sdk entities
